+++ backend/src/main/java/com/example/backend/servicies/CoordinatesService.java
@@ -3,7 +3,6 @@ package com.example.backend.servicies;
 import com.example.backend.entities.Coordinates;
 import com.example.backend.entities.DTO.CoordinatesDTO;
 import com.example.backend.repositories.CoordinatesRepository;
-import com.example.backend.utils.Validator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
@@ -19,19 +18,21 @@ public class CoordinatesService {
     private final CoordinatesRepository coordinatesRepository;
     private final UserService userService;
     private final SimpMessagingTemplate simpMessagingTemplate;
+    private final Checker checker;
 
     @Autowired
-    public CoordinatesService(CoordinatesRepository coordinatesRepository, UserService userService, SimpMessagingTemplate simpMessagingTemplate) {
+    public CoordinatesService(CoordinatesRepository coordinatesRepository, UserService userService, SimpMessagingTemplate simpMessagingTemplate, Checker checker) {
         this.coordinatesRepository = coordinatesRepository;
         this.userService = userService;
         this.simpMessagingTemplate = simpMessagingTemplate;
+        this.checker = checker;
     }
     public List<Coordinates> getAllCoordinates(){
         return coordinatesRepository.findAll();
     }
 
     public ResponseEntity<?> addCoordinate(CoordinatesDTO coordinatesDTO) {
-        ResponseEntity<?> resp = Validator.validate(coordinatesDTO);
+        ResponseEntity<?> resp = checker.validate(coordinatesDTO);
         if (resp.getStatusCode() != HttpStatus.OK) {
             return resp;
         }
@@ -49,13 +50,17 @@ public class CoordinatesService {
     }
 
     public ResponseEntity<?> updateCoordinates(Integer id, CoordinatesDTO coordinatesDTO) {
-        ResponseEntity<?> resp = Validator.validate(coordinatesDTO);
+        ResponseEntity<?> resp = checker.validate(coordinatesDTO);
         if (resp.getStatusCode() != HttpStatus.OK) {
             return resp;
         }
         Coordinates coordinates = coordinatesRepository.findById(id).orElse(null);
         if (coordinates == null) {
             return new ResponseEntity<>("Error: Coordinates not found", HttpStatus.NOT_FOUND);
+        }
+        resp = checker.check_rights(coordinates);
+        if (resp.getStatusCode() != HttpStatus.OK) {
+            return resp;
         }
         coordinates.setAuthor(userService.getCurrentUsername());
         coordinates.setX(coordinatesDTO.getX());
@@ -73,6 +78,10 @@ public class CoordinatesService {
         Coordinates coordinates = coordinatesRepository.findById(id).orElse(null);
         if (coordinates == null) {
             return new ResponseEntity<>("Error: Coordinates not found", HttpStatus.NOT_FOUND);
+        }
+        ResponseEntity<?> resp = checker.check_rights(coordinates);
+        if (resp.getStatusCode() != HttpStatus.OK) {
+            return resp;
         }
         coordinatesRepository.delete(coordinates);
         simpMessagingTemplate.convertAndSend("/topic/coordinates", getAllCoordinates());
